+++ src/sim/space_set.cc
@@ -109,7 +109,7 @@ void SpaceSet::remove(Object * obj)
 
 //------------------------------------------------------------------------------
 
-Object * SpaceSet::newObject(const ObjectTag tag, unsigned num)
+Object * SpaceSet::newObject(const ObjectTag tag, size_t num)
 {
     if ( tag == Space::TAG )
     {
